+++ packages/selection-plugin/src/lib/utils.spec.ts
@@ -1,6 +1,5 @@
-import { Component, NodeEditor } from '@naetverkjs/naetverk';
-import { Comp1, Comp2 } from '@naetverkjs/naetverk/test/components';
-import { renderMock } from '@naetverkjs/naetverk/test/utils/render-mock';
+import { NodeEditor } from '@naetverkjs/naetverk';
+import { initRenderer } from '@naetverkjs/naetverk/test/init-render.helper';
 import { Position } from './interfaces/position.interface';
 import {
   applyTransform,
@@ -59,27 +58,12 @@ describe('Until', () => {
   });
 
   describe('node selection on rectangle', () => {
-    let container: HTMLElement;
     let editor: NodeEditor;
-    let comps: Component[];
-
-    beforeEach(() => {
-      const par = document.createElement('div') as HTMLElement;
-      container = document.createElement('div') as HTMLElement;
-      par.appendChild(container);
-      editor = new NodeEditor('test@0.0.1', container);
-      comps = [new Comp1(), new Comp2()];
-      comps.forEach((c) => editor.register(c));
-    });
 
     it('should select a node', async () => {
-      renderMock(editor);
-      const node1 = await comps[0].createNode();
-      const node2 = await comps[0].createNode();
-      node1.position = [100, 100];
-      node2.position = [300, 300];
-      editor.addNode(node1);
-      editor.addNode(node2);
+      editor = await initRenderer({
+        nodes: [{ position: [100, 100] }, { position: [300, 300] }],
+      });
       expect(
         getNodesFromSelectionArea(editor, [
           { x: 0, y: 0 },
